python/requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.4.0
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.7
